+++ test/components/toolbar/toolbar-searchfield.e2e-spec.js
@@ -44,11 +44,11 @@ describe('Searchfield with Toolbar alignment tests', () => {
       await browser.driver
         .wait(protractor.ExpectedConditions.presenceOf(searchfieldInputEl), config.waitsFor);
       await browser.driver.sleep(config.sleep);
-      expect(await browser.imageComparison.checkElement(searchfieldInputEl, 'toolbar-searchfield-init')).toEqual(0);
+      expect(await browser.imageComparison.checkElement(searchfieldInputEl, 'toolbar-searchfield-init')).toBeLessThan(0.2);
 
       await searchfieldInputEl.click();
       await browser.driver.sleep(config.sleep);
-      expect(await browser.imageComparison.checkElement(searchfieldToolbarContainer, 'toolbar-searchfield-alignment')).toEqual(0);
+      expect(await browser.imageComparison.checkElement(searchfieldToolbarContainer, 'toolbar-searchfield-alignment')).toBeLessThan(0.2);
     });
   }
 });